+++ docs/dataset.rst
@@ -1,0 +1,7 @@
+Dataset
+=======
+
+Inherits from `pytorch datasets <https://pytorch.org/docs/stable/data.html#torch.utils.data.Dataset>`_ to allow use with `pytorch dataloader <https://pytorch.org/docs/stable/data.html#torch.utils.data.DataLoader>`_.
+
+.. automodule:: time_series_dataset
+  :members:
