+++ docs/metrics.rst
@@ -3,5 +3,5 @@ Metrics
 
 Common evaluation metrics for time series forecasts.
 
-.. automodule:: custom_metrics
+.. automodule:: metrics
   :members:
