+++ docs/transforms.rst
@@ -1,0 +1,7 @@
+Transformations
+===============
+
+Transformations of the time series intended to be used in a similar fashion to `torchvision <https://pytorch.org/docs/stable/torchvision/transforms.html>`_.
+
+.. automodule:: transforms
+  :members:
