--- docs/utils.rst
@@ -1,26 +1,0 @@
-Utils
-=====
-
-Time Series Dataset
--------------------
-
-.. autoclass:: utils.TimeSeriesDataset
-  :members:
-
-Transformations
----------------
-
-.. autoclass:: utils.LogTransform
-  :members:
-
-.. autoclass:: utils.RemoveLast
-  :members:
-
-.. autoclass:: utils.Standardize
-  :members:
-
-.. autoclass:: utils.Tensorize
-  :members:
-
-.. autoclass:: utils.Target
-  :members:
